refactor: split test_utils into a module tree
Break the single src/test_utils.rs into src/test_utils/{mod,raw_ssz,
steps,loader}.rs. Pure file organization: behavior, fixture semantics,
and dependencies are unchanged. Public names remain reachable at
crate::test_utils::* via re-exports; the Raw* items the loader reads
across the new module boundary become pub(crate).
