--- src/test_utils.rs
@@ -1,650 +1,0 @@
-//! # Test Utilities (Unstable)
-//!
-//! **Warning:** This module is not part of the stable public API.
-//! It may change or be removed without notice in any version.
-//!
-//! Provides helpers for loading Ethereum consensus-spec test fixtures,
-//! useful for integration testing and downstream library testing.
-//!
-//! Enable with the `test-utils` feature:
-//! ```toml
-//! [dev-dependencies]
-//! eth-light-client = { version = "0.1", features = ["test-utils"] }
-//! ```
-
-use crate::types::consensus::{
-    BeaconBlockHeader, ExecutionPayloadHeaderCapella, LightClientBootstrap,
-    LightClientHeader as PubLightClientHeader, LightClientUpdate, SyncAggregate, SyncCommittee,
-};
-use crate::types::primitives::Root;
-use crate::types::primitives::{Bloom, ExtraData};
-use ssz_rs::prelude::*;
-use std::fs;
-use std::path::{Path, PathBuf};
-
-/// Fork tag used by the fixture loader to wrap deserialized headers
-/// into the correct [`LightClientHeader`](PubLightClientHeader) variant.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TestFork {
-    Altair,
-    Bellatrix,
-    Capella,
-}
-
-impl TestFork {
-    /// Only valid for Altair/Bellatrix; Capella headers carry an execution payload (panics).
-    fn wrap_header(&self, beacon: BeaconBlockHeader) -> PubLightClientHeader {
-        match self {
-            TestFork::Altair => PubLightClientHeader::altair(beacon),
-            TestFork::Bellatrix => PubLightClientHeader::bellatrix(beacon),
-            TestFork::Capella => {
-                panic!("Capella headers require execution payload; use Capella-specific load path")
-            }
-        }
-    }
-
-    /// Return a `ChainSpec` whose fork schedule matches the spec-test fixtures
-    /// for this fork.
-    pub fn chain_spec(&self) -> crate::config::ChainSpec {
-        use crate::config::{ChainSpec, ChainSpecConfig};
-
-        let mut config = ChainSpecConfig {
-            genesis_time: 1578009600,
-            seconds_per_slot: 6,
-            slots_per_epoch: 8,
-            epochs_per_sync_committee_period: 8,
-            sync_committee_size: 32,
-            altair_fork_version: [0x01, 0x00, 0x00, 0x01],
-            altair_fork_epoch: 0,
-            bellatrix_fork_version: [0x02, 0x00, 0x00, 0x01],
-            bellatrix_fork_epoch: u64::MAX,
-            capella_fork_version: [0x03, 0x00, 0x00, 0x01],
-            capella_fork_epoch: u64::MAX,
-            deneb_fork_version: [0x04, 0x00, 0x00, 0x01],
-            deneb_fork_epoch: u64::MAX,
-            electra_fork_version: [0x05, 0x00, 0x00, 0x01],
-            electra_fork_epoch: u64::MAX,
-        };
-
-        match self {
-            TestFork::Altair => {} // defaults are correct
-            TestFork::Bellatrix => {
-                config.bellatrix_fork_epoch = 0;
-            }
-            TestFork::Capella => {
-                config.bellatrix_fork_epoch = 0;
-                config.capella_fork_epoch = 0;
-            }
-        }
-
-        ChainSpec::try_from_config(config).expect("minimal fixture config is valid")
-    }
-}
-
-// ============================================================================
-// SSZ Deserialization Types (internal)
-// ============================================================================
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawBeaconBlockHeader {
-    slot: u64,
-    proposer_index: u64,
-    parent_root: Node,
-    state_root: Node,
-    body_root: Node,
-}
-
-impl RawBeaconBlockHeader {
-    fn into_beacon_block_header(self) -> BeaconBlockHeader {
-        let mut parent_root = [0u8; 32];
-        parent_root.copy_from_slice(self.parent_root.as_ref());
-        let mut state_root = [0u8; 32];
-        state_root.copy_from_slice(self.state_root.as_ref());
-        let mut body_root = [0u8; 32];
-        body_root.copy_from_slice(self.body_root.as_ref());
-
-        BeaconBlockHeader::new(
-            self.slot,
-            self.proposer_index,
-            parent_root,
-            state_root,
-            body_root,
-        )
-    }
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct LightClientHeader {
-    beacon: RawBeaconBlockHeader,
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawLightClientBootstrap {
-    header: LightClientHeader,
-    current_sync_committee: RawSyncCommittee,
-    current_sync_committee_branch: Vector<Node, 5>,
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawSyncCommittee {
-    pubkeys: Vector<Vector<u8, 48>, 32>,
-    aggregate_pubkey: Vector<u8, 48>,
-}
-
-impl RawSyncCommittee {
-    fn to_sync_committee(&self) -> Result<SyncCommittee, String> {
-        if self.pubkeys.len() != 32 {
-            return Err(format!(
-                "Expected 32 pubkeys (minimal preset), got {}",
-                self.pubkeys.len()
-            ));
-        }
-
-        let mut pubkeys_array = Box::new([[0u8; 48]; 512]);
-        for (i, pk) in self.pubkeys.iter().enumerate() {
-            let mut key = [0u8; 48];
-            key.copy_from_slice(pk.as_ref());
-            pubkeys_array[i] = key;
-        }
-
-        let mut aggregate = [0u8; 48];
-        aggregate.copy_from_slice(self.aggregate_pubkey.as_ref());
-
-        Ok(SyncCommittee::new(pubkeys_array, aggregate))
-    }
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawSyncAggregate {
-    sync_committee_bits: Bitvector<32>,
-    sync_committee_signature: Vector<u8, 96>,
-}
-
-impl RawSyncAggregate {
-    fn into_sync_aggregate(self) -> Result<SyncAggregate, String> {
-        let mut bits_array = Box::new([false; 512]);
-        for (i, bit) in self.sync_committee_bits.iter().enumerate() {
-            bits_array[i] = *bit;
-        }
-
-        let mut signature = [0u8; 96];
-        signature.copy_from_slice(self.sync_committee_signature.as_ref());
-
-        Ok(SyncAggregate::new(bits_array, signature))
-    }
-}
-
-// Altair/Bellatrix update (beacon-only headers)
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawLightClientUpdate {
-    attested_header: LightClientHeader,
-    next_sync_committee: RawSyncCommittee,
-    next_sync_committee_branch: Vector<Node, 5>,
-    finalized_header: LightClientHeader,
-    finality_branch: Vector<Node, 6>,
-    sync_aggregate: RawSyncAggregate,
-    signature_slot: u64,
-}
-
-// ============================================================================
-// Capella SSZ Types (execution payload in headers)
-// ============================================================================
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawExecutionPayloadHeader {
-    parent_hash: Node,
-    fee_recipient: Vector<u8, 20>,
-    state_root: Node,
-    receipts_root: Node,
-    logs_bloom: Vector<u8, 256>,
-    prev_randao: Node,
-    block_number: u64,
-    gas_limit: u64,
-    gas_used: u64,
-    timestamp: u64,
-    extra_data: List<u8, 32>,
-    base_fee_per_gas: ssz_rs::U256,
-    block_hash: Node,
-    transactions_root: Node,
-    withdrawals_root: Node,
-}
-
-impl RawExecutionPayloadHeader {
-    fn into_execution_payload_header(self) -> Result<ExecutionPayloadHeaderCapella, String> {
-        fn node_to_root(n: &Node) -> [u8; 32] {
-            let mut r = [0u8; 32];
-            r.copy_from_slice(n.as_ref());
-            r
-        }
-
-        let mut fee_recipient = [0u8; 20];
-        fee_recipient.copy_from_slice(self.fee_recipient.as_ref());
-
-        let mut bloom_bytes = [0u8; 256];
-        bloom_bytes.copy_from_slice(self.logs_bloom.as_ref());
-
-        // Convert ssz_rs::U256 to ruint::U256 via LE bytes
-        let le_bytes = self.base_fee_per_gas.to_bytes_le();
-        let mut u256_bytes = [0u8; 32];
-        let len = le_bytes.len().min(32);
-        u256_bytes[..len].copy_from_slice(&le_bytes[..len]);
-        let base_fee = ruint::aliases::U256::from_le_bytes(u256_bytes);
-
-        let extra_data_vec: Vec<u8> = self.extra_data.to_vec();
-        let extra_data = ExtraData::try_new(extra_data_vec).map_err(|e| e.to_string())?;
-
-        Ok(ExecutionPayloadHeaderCapella {
-            parent_hash: node_to_root(&self.parent_hash),
-            fee_recipient,
-            state_root: node_to_root(&self.state_root),
-            receipts_root: node_to_root(&self.receipts_root),
-            logs_bloom: Bloom(bloom_bytes),
-            prev_randao: node_to_root(&self.prev_randao),
-            block_number: self.block_number,
-            gas_limit: self.gas_limit,
-            gas_used: self.gas_used,
-            timestamp: self.timestamp,
-            extra_data,
-            base_fee_per_gas: base_fee,
-            block_hash: node_to_root(&self.block_hash),
-            transactions_root: node_to_root(&self.transactions_root),
-            withdrawals_root: node_to_root(&self.withdrawals_root),
-        })
-    }
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawCapellaLightClientHeader {
-    beacon: RawBeaconBlockHeader,
-    execution: RawExecutionPayloadHeader,
-    execution_branch: Vector<Node, 4>,
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawCapellaLightClientBootstrap {
-    header: RawCapellaLightClientHeader,
-    current_sync_committee: RawSyncCommittee,
-    current_sync_committee_branch: Vector<Node, 5>,
-}
-
-#[derive(Debug, Clone, Default, SimpleSerialize)]
-struct RawCapellaLightClientUpdate {
-    attested_header: RawCapellaLightClientHeader,
-    next_sync_committee: RawSyncCommittee,
-    next_sync_committee_branch: Vector<Node, 5>,
-    finalized_header: RawCapellaLightClientHeader,
-    finality_branch: Vector<Node, 6>,
-    sync_aggregate: RawSyncAggregate,
-    signature_slot: u64,
-}
-
-fn raw_capella_header_to_pub(
-    raw: &RawCapellaLightClientHeader,
-) -> Result<PubLightClientHeader, String> {
-    let beacon = raw.beacon.clone().into_beacon_block_header();
-    let execution = raw.execution.clone().into_execution_payload_header()?;
-    let mut execution_branch = [[0u8; 32]; 4];
-    for (i, node) in raw.execution_branch.iter().enumerate() {
-        execution_branch[i].copy_from_slice(node.as_ref());
-    }
-    Ok(PubLightClientHeader::capella(
-        beacon,
-        execution,
-        execution_branch,
-    ))
-}
-
-fn raw_capella_update_to_pub(
-    raw: RawCapellaLightClientUpdate,
-) -> Result<LightClientUpdate, String> {
-    let sync_committee = raw.next_sync_committee.to_sync_committee()?;
-    let sync_aggregate = raw.sync_aggregate.into_sync_aggregate()?;
-
-    let has_sync_committee = !sync_committee
-        .pubkeys
-        .iter()
-        .all(|pk| pk.iter().all(|&b| b == 0));
-
-    let finality_branch: Vec<[u8; 32]> = raw
-        .finality_branch
-        .iter()
-        .map(|node| {
-            let mut root = [0u8; 32];
-            root.copy_from_slice(node.as_ref());
-            root
-        })
-        .collect();
-
-    let next_sync_committee_branch: Vec<[u8; 32]> = raw
-        .next_sync_committee_branch
-        .iter()
-        .map(|node| {
-            let mut root = [0u8; 32];
-            root.copy_from_slice(node.as_ref());
-            root
-        })
-        .collect();
-
-    // A default finalized header (slot=0) means no finality update.
-    let has_finality = raw.finalized_header.beacon.slot != 0;
-    let finalized_header = if has_finality {
-        Some(raw_capella_header_to_pub(&raw.finalized_header)?)
-    } else {
-        None
-    };
-
-    Ok(LightClientUpdate {
-        attested_header: raw_capella_header_to_pub(&raw.attested_header)?,
-        finalized_header,
-        finality_branch: if has_finality {
-            finality_branch
-        } else {
-            Vec::new()
-        },
-        next_sync_committee: if has_sync_committee {
-            Some(sync_committee)
-        } else {
-            None
-        },
-        next_sync_committee_branch: if has_sync_committee {
-            next_sync_committee_branch
-        } else {
-            Vec::new()
-        },
-        sync_aggregate,
-        signature_slot: raw.signature_slot,
-    })
-}
-
-impl RawLightClientUpdate {
-    fn into_light_client_update(self, fork: TestFork) -> Result<LightClientUpdate, String> {
-        let sync_committee = self.next_sync_committee.to_sync_committee()?;
-        let sync_aggregate = self.sync_aggregate.into_sync_aggregate()?;
-
-        let has_sync_committee = !sync_committee
-            .pubkeys
-            .iter()
-            .all(|pk| pk.iter().all(|&b| b == 0));
-
-        let finality_branch: Vec<Root> = self
-            .finality_branch
-            .iter()
-            .map(|node| {
-                let mut root = [0u8; 32];
-                root.copy_from_slice(node.as_ref());
-                root
-            })
-            .collect();
-
-        let next_sync_committee_branch: Vec<Root> = self
-            .next_sync_committee_branch
-            .iter()
-            .map(|node| {
-                let mut root = [0u8; 32];
-                root.copy_from_slice(node.as_ref());
-                root
-            })
-            .collect();
-
-        Ok(LightClientUpdate {
-            attested_header: fork
-                .wrap_header(self.attested_header.beacon.into_beacon_block_header()),
-            finalized_header: Some(
-                fork.wrap_header(self.finalized_header.beacon.into_beacon_block_header()),
-            ),
-            finality_branch,
-            next_sync_committee: if has_sync_committee {
-                Some(sync_committee)
-            } else {
-                None
-            },
-            next_sync_committee_branch: if has_sync_committee {
-                next_sync_committee_branch
-            } else {
-                Vec::new()
-            },
-            sync_aggregate,
-            signature_slot: self.signature_slot,
-        })
-    }
-}
-
-// ============================================================================
-// Test Metadata Types (YAML)
-// ============================================================================
-
-/// Metadata from a spec test's meta.yaml file.
-#[derive(Debug, serde::Deserialize)]
-pub struct TestMeta {
-    pub genesis_validators_root: String,
-    #[allow(dead_code)]
-    trusted_block_root: String,
-    #[allow(dead_code)]
-    bootstrap_fork_digest: String,
-    #[allow(dead_code)]
-    store_fork_digest: String,
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct StateChecks {
-    pub finalized_header: Option<HeaderCheck>,
-    pub optimistic_header: Option<HeaderCheck>,
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct HeaderCheck {
-    pub slot: u64,
-    pub beacon_root: String,
-    /// Present only for Capella+ (absent for Altair/Bellatrix).
-    #[serde(default)]
-    pub execution_root: Option<String>,
-}
-
-/// A single test step from steps.yaml.
-#[derive(Debug, serde::Deserialize)]
-#[serde(untagged)]
-pub enum TestStep {
-    ProcessUpdate {
-        process_update: ProcessUpdateStep,
-    },
-    /// Force update (safety timeout mechanism).
-    ForceUpdate {
-        force_update: ForceUpdateStep,
-    },
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct ProcessUpdateStep {
-    #[allow(dead_code)]
-    update_fork_digest: String,
-    /// Update file name (without .ssz_snappy extension).
-    pub update: String,
-    pub current_slot: u64,
-    pub checks: StateChecks,
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct ForceUpdateStep {
-    pub current_slot: u64,
-    pub checks: StateChecks,
-}
-
-// ============================================================================
-// Public API
-// ============================================================================
-
-#[derive(Debug, Clone)]
-pub struct BootstrapData {
-    pub header: PubLightClientHeader,
-    pub sync_committee: SyncCommittee,
-    pub branch: Vec<Root>,
-    pub genesis_validators_root: Root,
-}
-
-impl BootstrapData {
-    pub fn into_bootstrap(self) -> LightClientBootstrap {
-        LightClientBootstrap::from_header(
-            self.header,
-            self.sync_committee,
-            self.branch,
-            self.genesis_validators_root,
-        )
-    }
-}
-
-/// Loads spec test fixtures from a directory.
-///
-/// **Unstable:** This API may change without notice.
-pub struct SpecTestLoader {
-    test_dir: PathBuf,
-    fork: TestFork,
-}
-
-impl SpecTestLoader {
-    /// Loader for the minimal Altair light-client sync fixtures.
-    pub fn minimal_altair_sync() -> Self {
-        let test_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("tests/fixtures/minimal/altair/light_client/sync/light_client_sync");
-        Self {
-            test_dir,
-            fork: TestFork::Altair,
-        }
-    }
-
-    /// Loader for the minimal Bellatrix light-client sync fixtures.
-    pub fn minimal_bellatrix_sync() -> Self {
-        let test_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("tests/fixtures/minimal/bellatrix/light_client/sync/light_client_sync");
-        Self {
-            test_dir,
-            fork: TestFork::Bellatrix,
-        }
-    }
-
-    /// Loader for the minimal Capella light-client sync fixtures.
-    pub fn minimal_capella_sync() -> Self {
-        let test_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("tests/fixtures/minimal/capella/light_client/sync/light_client_sync");
-        Self {
-            test_dir,
-            fork: TestFork::Capella,
-        }
-    }
-
-    pub fn from_path(path: impl Into<PathBuf>, fork: TestFork) -> Self {
-        Self {
-            test_dir: path.into(),
-            fork,
-        }
-    }
-
-    pub fn chain_spec(&self) -> crate::config::ChainSpec {
-        self.fork.chain_spec()
-    }
-
-    pub fn load_bootstrap(&self) -> Result<BootstrapData, Box<dyn std::error::Error>> {
-        let meta = self.load_meta()?;
-        let genesis_validators_root = hex_to_root(&meta.genesis_validators_root)?;
-        let bootstrap_path = self.test_dir.join("bootstrap.ssz_snappy");
-
-        match self.fork {
-            TestFork::Altair | TestFork::Bellatrix => {
-                let bootstrap: RawLightClientBootstrap = load_ssz_snappy(&bootstrap_path)?;
-                let sync_committee = bootstrap.current_sync_committee.to_sync_committee()?;
-                let branch = nodes_to_roots(&bootstrap.current_sync_committee_branch);
-
-                Ok(BootstrapData {
-                    header: self
-                        .fork
-                        .wrap_header(bootstrap.header.beacon.into_beacon_block_header()),
-                    sync_committee,
-                    branch,
-                    genesis_validators_root,
-                })
-            }
-            TestFork::Capella => {
-                let bootstrap: RawCapellaLightClientBootstrap = load_ssz_snappy(&bootstrap_path)?;
-                let sync_committee = bootstrap.current_sync_committee.to_sync_committee()?;
-                let branch = nodes_to_roots(&bootstrap.current_sync_committee_branch);
-                let header = raw_capella_header_to_pub(&bootstrap.header)?;
-
-                Ok(BootstrapData {
-                    header,
-                    sync_committee,
-                    branch,
-                    genesis_validators_root,
-                })
-            }
-        }
-    }
-
-    /// `name` must not include the `.ssz_snappy` extension.
-    pub fn load_update(&self, name: &str) -> Result<LightClientUpdate, Box<dyn std::error::Error>> {
-        let update_path = self.test_dir.join(format!("{}.ssz_snappy", name));
-
-        match self.fork {
-            TestFork::Altair | TestFork::Bellatrix => {
-                let raw: RawLightClientUpdate = load_ssz_snappy(&update_path)?;
-                raw.into_light_client_update(self.fork)
-                    .map_err(|e| e.into())
-            }
-            TestFork::Capella => {
-                let raw: RawCapellaLightClientUpdate = load_ssz_snappy(&update_path)?;
-                raw_capella_update_to_pub(raw).map_err(|e| e.into())
-            }
-        }
-    }
-
-    pub fn load_meta(&self) -> Result<TestMeta, Box<dyn std::error::Error>> {
-        let meta_path = self.test_dir.join("meta.yaml");
-        let meta_contents = fs::read_to_string(&meta_path)?;
-        let meta: TestMeta = serde_yaml::from_str(&meta_contents)?;
-        Ok(meta)
-    }
-
-    pub fn load_steps(&self) -> Result<Vec<TestStep>, Box<dyn std::error::Error>> {
-        let steps_path = self.test_dir.join("steps.yaml");
-        let steps_contents = fs::read_to_string(&steps_path)?;
-        let steps: Vec<TestStep> = serde_yaml::from_str(&steps_contents)?;
-        Ok(steps)
-    }
-}
-
-// ============================================================================
-// Internal Helpers
-// ============================================================================
-
-fn nodes_to_roots(nodes: &[Node]) -> Vec<Root> {
-    nodes
-        .iter()
-        .map(|node| {
-            let mut root = [0u8; 32];
-            root.copy_from_slice(node.as_ref());
-            root
-        })
-        .collect()
-}
-
-fn load_ssz_snappy<T>(file_path: &Path) -> Result<T, Box<dyn std::error::Error>>
-where
-    T: Deserialize,
-{
-    let compressed = fs::read(file_path)?;
-    let mut decoder = snap::raw::Decoder::new();
-    let decompressed = decoder.decompress_vec(&compressed)?;
-    let decoded =
-        T::deserialize(&decompressed).map_err(|e| format!("SSZ decode error: {:?}", e))?;
-    Ok(decoded)
-}
-
-/// Convert a hex string (with or without 0x prefix) to a 32-byte root.
-pub fn hex_to_root(hex: &str) -> Result<Root, Box<dyn std::error::Error>> {
-    let hex = hex.strip_prefix("0x").unwrap_or(hex);
-    let bytes = hex::decode(hex)?;
-    if bytes.len() != 32 {
-        return Err(format!("Expected 32 bytes, got {}", bytes.len()).into());
-    }
-    let mut root = [0u8; 32];
-    root.copy_from_slice(&bytes);
-    Ok(root)
-}
